:bug: Validate that file arguments are not empty in CLI commands
- Add NoFilesError for empty files validation in lint, tidy, and ids commands
- Consolidate CLI error classes into cli.rb
- Remove separate error class files
