--- lib/foxtail/cli/error.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Foxtail
-  module CLI
-    # Base error class for CLI-specific exceptions
-    class Error < Foxtail::Error; end
-  end
-end
